requirements.txt
fastapi
uvicorn
requests
chromadb
sentence-transformers
PyMuPDF
python-docx
python-dotenv
